third_party/blink/web_tests/platform/generic/external/wpt/service-workers/service-worker/fetch-event-is-reload-iframe-navigation-manual.https-expected.txt

This is a testharness.js-based test.
FAIL FetchEvent#request.isReloadNavigation is true for manual reload. assert_equals: expected "method = GET, isReloadNavigation = false" but got "method = GET, isReloadNavigation = undefined"
Harness: the test ran to completion.
